Do not display map when lon lat is null
